pyproject.toml
[tool.poetry]
name = "caluma"
version = "8.0.0"
description = "Caluma Service providing GraphQL API"
homepage = "https://caluma.io"
repository = "https://github.com/projectcaluma/caluma"
documentation = "https://caluma.gitbook.io/"
authors = ["Caluma <info@caluma.io>"]
license = "GPL-3.0-or-later"
readme = "README.md"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3.8",
]
include = ["CHANGELOG.md"]
exclude = [
    "caluma/**/tests",
]

[tool.poetry.dependencies]
python = "^3.8.1"
dateparser = "^1.1.0"
Django = "^3.2.19"
django-cors-headers = "^3.11.0"
django-environ = "^0.10.0"
django-filter = "^23.1"
django-localized-fields = "^6.6"
django-postgres-extra = "^2.0.4"
django-watchman = "^1.2.0"
djangorestframework = "^3.13.1"
django-simple-history = "^3.0.0"
graphene-django = "3.0.0b7"
graphql-relay = "^3.1.5"
idna = "^3.3"
minio = "^7.1.4"
psycopg2-binary = "^2.9.3"
pyjexl = "^0.3.0"
python-memcached = "^1.59"
requests = "^2.27.1"
urllib3 = "^1.26.8"
uWSGI = "^2.0.20"
graphql-core = "~3.1.7"

[tool.poetry.group.dev.dependencies]
# Dev dependencies are always pinned to an exact version (No caret or tilde
# matching) to ensure everyone uses the same version of the tools, especially
# linters & formatters
black = "23.3.0"
django-extensions = "3.2.1"
factory-boy = "3.2.1"
Faker = "18.7.0"
flake8 = "6.0.0"
flake8-bugbear = "23.5.9"
flake8-debugger = "4.1.2"
flake8-docstrings = "1.7.0"
flake8-isort = "6.0.0"
flake8-string-format = "0.3.0"
flake8-tuple = "0.4.1"
gitlint = "0.19.1"
isort = "5.12.0"
pdbpp = "0.10.3"
pre-commit = "3.3.1"
pydocstyle = "6.3.0"
pytest = "7.3.1"
pytest-cov = "4.0.0"
pytest-django = "4.5.2"
pytest-env = "0.8.1"
pytest-factoryboy = "2.5.1"
pytest-freezegun = "0.4.2"
pytest-mock = "3.10.0"
pytest-randomly = "3.12.0"
pytest-xdist = "3.2.1"
python-semantic-release = "7.33.3"
requests-mock = "1.10.0"
reuse = "1.1.2"
syrupy = "4.0.2"

[tool.isort]
skip = [
  "migrations",
  "snapshots",
]
known_first_party = ["caluma"]
known_third_party = ["dateparser"]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true
line_length = 88

[tool.pytest.ini_options]
addopts = "-n auto --reuse-db --randomly-seed=1521188766 --randomly-dont-reorganize"
DJANGO_SETTINGS_MODULE = "caluma.settings.django"
env = [
    "META_FIELDS=test-key,foobar",
    "OIDC_USERINFO_ENDPOINT=mock://caluma.io/openid/userinfo",
    "OIDC_BEARER_TOKEN_REVALIDATION_TIME=60",
    "LANGUAGES=en,de,fr",
    "ENABLE_HISTORICAL_API=true",
    "ENABLE_ACCESS_LOG=true",
    "ENABLE_HEALTHZ_ENDPOINT=true",
]
filterwarnings = [
    "error::DeprecationWarning",
    "error::PendingDeprecationWarning",
    "ignore:The 'arrayconnection' module is deprecated:DeprecationWarning", # deprecation in graphene
    "ignore:distutils Version classes are deprecated:DeprecationWarning", # deprecation in pytest-freezegun
    "ignore:'django_extensions' defines default_app_config:PendingDeprecationWarning", # deprecation in django_extensions
    "ignore::requests.packages.urllib3.exceptions.InsecureRequestWarning", # MinIO tests do "insecure" requests - that's ok
    "ignore:invalid escape sequence",
    "ignore:pkg_resources is deprecated as an API:DeprecationWarning", # deprecation in django-simple-history
]

[tool.coverage.run]
source = ["."]

[tool.coverage.report]
fail_under = 100
exclude_lines = [
    "pragma: no cover",
    "pragma: todo cover",
    "def __str__",
    "def __unicode__",
    "def __repr__",
]
omit = [
    "*/migrations/*",
    "*/apps.py",
    "manage.py",
    "caluma/settings_*.py",
    "caluma/wsgi.py",
]
show_missing = true

[tool.semantic_release]
version_source = "tag"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
